convert tests to kotlin, and move MimeType
