+++ vimeo-networking/src/main/java/com/vimeo/networking/model/VideoFile.java
@@ -41,29 +41,6 @@ import java.util.Date;
 @UseStag(FieldOption.SERIALIZED_NAME)
 public abstract class VideoFile implements Serializable {
 
-    @UseStag
-    public enum MimeType {
-        NONE("None"),
-        @SerializedName("video/mp4")
-        MP4("video/mp4"),
-        @SerializedName("video/webm")
-        WEBM("video/webm"), // Flash
-        @SerializedName("vp6/x-video")
-        VP6("vp6/x-video"); // Flash
-
-        private final String mTypeName;
-
-        MimeType(String typeName) {
-            mTypeName = typeName;
-        }
-
-        @Override
-        // Overridden for analytics.
-        public String toString() {
-            return mTypeName;
-        }
-    }
-
     public enum VideoQuality {
         NONE("N/A"),
         DASH("dash"),
